WIP: configuring SSR with Mike
