--- 00Base-Template/src/client/index.html
@@ -1,18 +1,0 @@
-<!doctype html>
-<html>
-	<head>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<meta name="keywords" content="">
-		<meta name="description" content="">
-		<meta name="author" content="">
-
-		<title>Base Template</title>
-
-		<link rel="stylesheet" type="text/css" href="/static/bundle.css">
-	</head>
-	<body id="body">
-		<div class="root"></div>
-		<script src="/static/bundle.js"></script>
-	</body>
-</html>
