+++ 00Base-Template/src/server/middleware/index.js
@@ -1,0 +1,2 @@
+export { default as handleRender } from './handleRender'
+export { default as assets } from './assets'
